==> 2. Elastic Net Search.ipynb <==
## Elastic Search

%%capture

!pip install elasticsearch==7.14.0
!apt install default-jdk > /dev/null

import os
import elasticsearch
from elasticsearch import Elasticsearch
import sys
import json
from ast import literal_eval
from tqdm import tqdm
import datetime
from elasticsearch import helpers

# Download & extract Elasticsearch 7.0.0

!wget https://artifacts.elastic.co/downloads/elasticsearch/elasticsearch-7.0.0-linux-x86_64.tar.gz -q
!tar -xzf elasticsearch-7.0.0-linux-x86_64.tar.gz
!chown -R daemon:daemon elasticsearch-7.0.0

# Creating daemon instance of elasticsearch
from subprocess import Popen, PIPE, STDOUT
es_server = Popen(['elasticsearch-7.0.0/bin/elasticsearch'],
                  stdout=PIPE, stderr=STDOUT,
                  preexec_fn=lambda: os.setuid(1)  # as daemon
                 )

# This part is important, since it takes a little amount of time for instance to load
import time
time.sleep(60)

%%bash
# If you get 1 root & 2 daemon process then Elasticsearch instance has started successfully
ps -ef | grep elasticsearch

# Check if elasticsearch is running
!curl -sX GET "localhost:9200/"

es = Elasticsearch(hosts = [{"host":"localhost", "port":9200}])
# Check if python is connected to elasticsearch
es.ping()

Settings = {
"mappings":{
    "properties":{
        "article":{
            "type":"text",
            'similarity': 'BM25'
        }
    }
}}

def json_formatter(dataset, index_name, index_type='_doc'):
    """
    This function is used to create JSON formatted dictionaries for Elasticsearch.

    Args:
      dataset: The dataset you want to apply this function.
      index_name: Name of the index in Elasticsearch
      index_type: Type of the index in Elasticsearch.
      Note: It is suggested to keep index_type as '_doc' since it is deprecated from version 6.
      Note: This function formats all columns of your dataset, if you want to apply this to special columns only,
      you can delete the second for loop and add your custom fields.
    """
    try:
        List = []

        for idx, row in enumerate(dataset):
            dic = {}
            dic['_index'] = index_name
            dic['sentence'] = row
            List.append(dic)
        return List

    except Exception as e:
        print("There is a problem: {}".format(e))

MY_INDEX = es.indices.create(index="corpus_arc", ignore=[400,404], body=Settings)
MY_INDEX

MY_INDEX = es.indices.create(index="corpus_aristo", ignore=[400,404], body=Settings)
MY_INDEX

index_exists = es.indices.exists(index="corpus_arc")

# Print the result
if index_exists:
    print(f"The index corpus_arc exists.")
else:
    print(f"The index corpus_arc does not exist.")

arc_json_formatted = json_formatter(dataset=corpus_arc, index_name='corpus_arc', index_type='_doc')

aristo_json_formatted = json_formatter(dataset=corpus_aristo, index_name='corpus_aristo', index_type='_doc')


res_arc = helpers.bulk(es, arc_json_formatted)
res_aristo= helpers.bulk(es, aristo_json_formatted)

output = pd.json_normalize((response['hits']['hits']))
output
