==> 4. Document Relevance Discriminator(DRD).ipynb <==
# Document Relevance Discriminator (DRD)

import os
import pandas as pd
import json
import numpy as np
import pickle
import tensorflow as tf
import transformers
from transformers import AutoTokenizer, TFT5ForConditionalGeneration
from tensorflow.keras.layers import Input, Dense, Activation, Dropout, LSTM, BatchNormalization
from tensorflow.keras.models import Model
from tensorflow.keras.callbacks import LearningRateScheduler, ReduceLROnPlateau, EarlyStopping, ModelCheckpoint
import datetime

### Loading SQUAD Dataset

# downoading SQUAD 2.0 dataset
!wget https://rajpurkar.github.io/SQuAD-explorer/dataset/train-v2.0.json
!wget https://rajpurkar.github.io/SQuAD-explorer/dataset/dev-v2.0.json


def load_squad(location):
# Load the SQuAD JSON dataset into a Python dictionary
    with open(location, 'r', encoding='utf-8') as file:
        squad_data = json.load(file)

    # Initialize lists to store data
    questions = []
    contexts = []
    labels = []

    # Process each example in the dataset
    for example in squad_data["data"]:
        for paragraph in example["paragraphs"]:
            context = paragraph["context"]
            for qa in paragraph["qas"]:
                question = qa["question"]
                is_impossible = qa.get("is_impossible", False)

                # Assign labels (0 for impossible, 1 for answerable)
                label = 0 if is_impossible else 1

                questions.append(question)
                contexts.append(context)
                labels.append(label)

    # Create a DataFrame
    df = pd.DataFrame({"question": questions, "context": contexts, "label": labels})

    return df

squad_train = load_squad("/content/train-v2.0.json" )
squad_dev = load_squad("/content/dev-v2.0.json")

import re

def remove_special_characters(input_string):
    # Use a regular expression to replace all non-alphanumeric characters with an empty string
    clean_string=[]
    for text in input_string:
        clean_string.append(re.sub(r'[^a-zA-Z0-9\s]', '', text))
    return clean_string

squad_train['context'] = remove_special_characters(squad_train['context'])
squad_dev['context'] = remove_special_characters(squad_dev['context'])

squad_train.head(3)

###training model

!pip install transformers

import tensorflow as tf
import transformers
from transformers import AutoTokenizer, TFT5ForConditionalGeneration
from tensorflow.keras.layers import Input, Dense, Activation, Dropout, LSTM, BatchNormalization
from tensorflow.keras.models import Model
from tensorflow.keras.callbacks import LearningRateScheduler, ReduceLROnPlateau, EarlyStopping, ModelCheckpoint
import datetime
import os

model_name = "google/flan-t5-large"
tokenizer = AutoTokenizer.from_pretrained(model_name)

# Calculate percentiles
percentiles = np.percentile([len(i.split())for i in squad_train['context']] , range(90, 101))

# Print the percentiles
for p, percentile_value in zip(range(90, 101), percentiles):
    print(f"{p}th percentile: {percentile_value}")

# Calculate percentiles
percentiles = np.percentile([len(i.split())for i in squad_train['question']] , range(90, 101))

# Print the percentiles
for p, percentile_value in zip(range(90, 101), percentiles):
    print(f"{p}th percentile: {percentile_value}")

from tqdm import tqdm

def tokenizing_df(dataset):

    max_seq_length = 512
    dataset_len =len(dataset)

    input_ids = np.zeros((dataset_len, max_seq_length))
    attention_mask= np.zeros((dataset_len, max_seq_length))

    target_input_ids= np.zeros((dataset_len, 2))
    target_attention_mask= np.zeros((dataset_len, 2))

    for i, row in tqdm(dataset.iterrows()):


        # now truncating tokens to max length

        # adding [CLS] at the begining and [SEP] at the end of the sentence
        pair = row['question'] + '</s>' + row['context']
        # converting tokens to unique IDs
        tokens = tokenizer.encode(pair)
        # adding zeros in the last if the size of the text is less than max length
        input_ids[i, :]= np.array(tokens + [0]*(max_seq_length-len(tokens)))[None,:512]
        # masking vector
        attention_mask[i, :]= np.array([1]*len(tokens) + [0]*(max_seq_length-len(tokens)))[None,:512]


        if row['label'] == 0:

          tokens = tokenizer.encode('A')

          target_input_ids[i, :] = np.array(tokens)[None,:]

          target_attention_mask[i, :] = np.array([1]*len(tokens))[None, :]

        else:

          tokens = tokenizer.encode('B')


          target_input_ids[i, :] = np.array(tokens)[None,:]

          target_attention_mask[i, :] = np.array([1]*len(tokens))[None, :]



    tokenized_data= np.array([input_ids, attention_mask],  dtype= 'int16')
    target_data =np.array([target_input_ids, target_attention_mask], dtype= 'int16')


    return tokenized_data, target_data

squad_x_train, squad_y_train = tokenizing_df(squad_train)
squad_x_dev, squad_y_dev = tokenizing_df(squad_dev)

pickle.dump((squad_x_train, squad_y_train, squad_x_dev, squad_y_dev),
            open('/content/drive/MyDrive/my assignments/33. A12 Reasoning Challenge- Self case study 2/squad_tokenized_t5_xl_512.pkl','wb'))

squad_x_train, squad_y_train, squad_x_dev, squad_y_dev= \
pickle.load(open('/content/drive/MyDrive/my assignments/33. A12 Reasoning Challenge- Self case study 2/squad_tokenized_t5_xl_512.pkl', 'rb'))

resolver = tf.distribute.cluster_resolver.TPUClusterResolver(tpu='')
tf.config.experimental_connect_to_cluster(resolver)
# This is the TPU initialization code that has to be at the beginning.
tf.tpu.experimental.initialize_tpu_system(resolver)
print("All devices: ", tf.config.list_logical_devices('TPU'))

from transformers import TFT5ForConditionalGeneration
strategy = tf.distribute.TPUStrategy(resolver)

optimizer = tf.keras.optimizers.AdamW(3e-5, clipnorm= 1.0, epsilon= 1e-6)
#loss= tf.keras.losses.BinaryCrossentropy()

with strategy.scope():
    model = TFT5ForConditionalGeneration.from_pretrained(model_name)
    model.compile(optimizer= optimizer,
                  metrics = ['accuracy'])

# function to save new best model with higher Validation accuracy
filepath="/content/drive/MyDrive/my assignments/33. A12 Reasoning Challenge- Self case study 2/DRD model on Squad/weights-{epoch:02d}-{val_accuracy:.4f}.tf"
checkpoint=  ModelCheckpoint(filepath=filepath,
                             monitor='val_accuracy',
                             save_best_only=True,
                             verbose=1,
                             mode='max',
                             save_weights_only=False,
                             options= tf.saved_model.SaveOptions(experimental_io_device="/job:localhost"))

# function to reduce learning rate if validation accuracy stop moving or decreasing
rlr = ReduceLROnPlateau(monitor="val_accuracy", factor= 0.1, min_lr= 0.000001,
           patience=1, verbose=1, mode = 'max')

# it will monitor validation accuracy if it has stop improving or giving constant accuracy at each epoch
# earlystop = EarlyStopping(monitor="val_accuracy", patience=2, verbose=1)

train = {'input_ids':squad_x_train[0], 'attention_mask': squad_x_train[1],
               'labels':squad_y_train[0], 'decoder_attention_mask':squad_y_train[1]}

dev = {'input_ids':squad_x_dev[0], 'attention_mask': squad_x_dev[1],
               'labels':squad_y_dev[0], 'decoder_attention_mask':squad_y_dev[1]}

epochs= 3
batch = 32

model.fit(x= train,
          validation_data= dev,
          verbose= 1,
          epochs= epochs,
          batch_size= batch,
          shuffle =True,)
        #   callbacks=[checkpoint, rlr])

model.save_weights('squad_saved_weights_t5_large.h5',
           options=  tf.saved_model.SaveOptions(experimental_io_device="/job:localhost"))

## Inference

from transformers import TFT5ForConditionalGeneration
optimizer = tf.keras.optimizers.AdamW(3e-5, clipnorm= 1.0, epsilon= 1e-6)
#loss= tf.keras.losses.BinaryCrossentropy()

model = TFT5ForConditionalGeneration.from_pretrained('google/flan-t5-large')
model.compile(optimizer= optimizer,
              metrics = ['accuracy'])

path= '/content/drive/MyDrive/my assignments/33. A12 Reasoning Challenge- Self case study 2/squad_saved_weights_t5_large.h5'
model.load_weights(path)
# ,  options = tf.train.CheckpointOptions(experimental_io_device="/job:localhost"))

# Accuracy score on dev dataset
from sklearn.metrics import accuracy_score
from tqdm import tqdm

squad_predicted= []
for i in tqdm(range(0, len(squad_x_dev[0]), 50)):
    squad_predicted.append(model.generate(squad_x_dev[0][i:i+50], output_scores=True, return_dict_in_generate=True))


squad_predicted_labels= [0 if tokenizer.decode(sequence[1]) == 'A' else 1 for i in squad_predicted for sequence in i['sequences']]

# Calculate the accuracy score
accuracy = accuracy_score(squad_dev['label'], squad_predicted_labels)

# Print the accuracy score
print(f"Accuracy: {accuracy}")

### generating DRD scores

from sklearn.preprocessing import MinMaxScaler
from tqdm import tqdm
from itertools import chain
from tensorflow.python.ops.numpy_ops import np_config
np_config.enable_numpy_behavior()

def drd_score(data):
  '''
  function to get DRD scores from downstream model trained on SQUAD dataset
  '''
    question= data['only_questions']
    answer= data['only_answers']
    context = data['context']

    scores = []
    # it will iterate through each question
    for i, que in enumerate(tqdm(question)):
        list_of_scores= []
        # it will iterate through each answer of that particular question
        # each question has four option so it will iterate four times
        for j, ans in enumerate(answer[i]):
            # it will get contexts of particular question with particular option
            # it will get 50 contexts
            context_len = len(context[i][j])
            # we will multiply question 50 times to match with context length
            pair= [que] * context_len

            # now we have 50 pairs of question and 50 contexts
            # now we will tokenize them
            tokens = tokenizer(pair, context[i][j], padding= True, truncation=True, return_tensors="tf")
            # model will generate probabilty scores of each 50 contexts

            predict = model.generate(tokens['input_ids'], output_scores=True, return_dict_in_generate=True)

            softmax = tf.nn.softmax(predict['scores'][0], axis= -1)

            max_values = np.max(softmax, axis = -1)

            # softmax = tf.nn.softmax(max_values, axis= -1)
            # softmax = [tf.nn.softmax(i).tolist() for i in max_values]


            list_of_scores.append(max_values.tolist())

        # now we will append softmax scores to the list of all 50 scores
        # each question have 4 options and each option has 50 context scores
        scores.append(list_of_scores)

    return scores

def joining_scores(data, new_scores):
  '''
  This fuction will join DRD scores with existing BM25 scores
  '''
    old_scores = data['score']
    scores =[]
    for idx_1, score_1 in enumerate(old_scores):
        list_of_scores= []
        for idx_2, score_2 in enumerate(score_1):

            list_of_scores.append([[score_3] + [new_scores[idx_1][idx_2][idx_3]] for idx_3, score_3 in enumerate(score_2)])
        scores.append(list_of_scores)

    return scores

train_challenge['score']= joining_scores(train_challenge, drd_score(train_challenge))
dev_challenge['score']= joining_scores(dev_challenge, drd_score(dev_challenge))
test_challenge['score']= joining_scores(test_challenge, drd_score(test_challenge))

pickle.dump((train_challenge, dev_challenge, test_challenge),
            open('/content/drive/MyDrive/my assignments/33. A12 Reasoning Challenge- Self case study 2/arc_datasets_scores_challenge.pkl','wb'))

train_easy['score']= joining_scores(train_easy, drd_score(train_easy))
dev_easy['score']= joining_scores(dev_easy, drd_score(dev_easy))
test_easy['score']= joining_scores(test_easy, drd_score(test_easy))

pickle.dump((train_easy, dev_easy, test_easy),
            open('/content/drive/MyDrive/my assignments/33. A12 Reasoning Challenge- Self case study 2/arc_datasets_scores_easy.pkl','wb'))

train_challenge, dev_challenge, test_challenge= \
pickle.load(open('/content/drive/MyDrive/my assignments/33. A12 Reasoning Challenge- Self case study 2/arc_datasets_scores_challenge.pkl', 'rb'))

train_easy, dev_easy, test_easy= \
pickle.load(open('/content/drive/MyDrive/my assignments/33. A12 Reasoning Challenge- Self case study 2/arc_datasets_scores_easy.pkl', 'rb'))
